requirements.txt
configargparse==1.5.3
lightning-bolts==0.7.0
lightning-lite==1.8.6
matplotlib==3.5.2
pandas==1.4.3
pytorch-lightning==1.9.5
rich==13.5.2
timm==0.9.12
torchmetrics==0.11.4
transformers==4.39.2
wandb==0.15.9
wilds==2.0.0
wget==3.2
